Start doing color parsing manually to avoid conversion bugs
This is related to a small issue in go-colorful which converts to and from
floats a number of times which can cause some off by ones in some of the colors.
A bug has been submitted upstream, but this is the simplest fix and removes a
dependency on an external library.

Fixes #4